+++ source/gear/firecracker-base.js
@@ -20,6 +20,6 @@ module.exports = new GearTemplate("Firecracker",
 		return dealDamage(targets, user, pendingDamage, false, element, adventure);
 	}
 ).setTargetingTags({ type: `random${SAFE_DELIMITER}3`, team: "foe", needsLivingTargets: true })
-	.setUpgrades("Double Firecracker", "Mercurial Firecracker", "Toxic Firecracker")
+	.setUpgrades("Double Firecracker", "Midas's Firecracker", "Toxic Firecracker")
 	.setDurability(15)
 	.setDamage(15);