+++ source/gear/firecracker-double.js
@@ -20,6 +20,6 @@ module.exports = new GearTemplate("Double Firecracker",
 		return dealDamage(targets, user, pendingDamage, false, element, adventure);
 	}
 ).setTargetingTags({ type: `random${SAFE_DELIMITER}6`, team: "foe", needsLivingTargets: true })
-	.setSidegrades("Mercurial Firecracker", "Toxic Firecracker")
+	.setSidegrades("Midas's Firecracker", "Toxic Firecracker")
 	.setDurability(15)
 	.setDamage(15);