+++ source/gear/firecracker-toxic.js
@@ -1,7 +1,7 @@
 const { GearTemplate } = require('../classes');
 const { SAFE_DELIMITER } = require('../constants.js');
 const { dealDamage, addModifier, changeStagger, getNames } = require('../util/combatantUtil.js');
-const { listifyEN, joinAsStatement } = require('../util/textUtil.js');
+const { joinAsStatement } = require('../util/textUtil.js');
 
 module.exports = new GearTemplate("Toxic Firecracker",
 	"Strike 3 random foes applying @{mod0Stacks} @{mod0} and @{damage} @{element} damage",
@@ -22,7 +22,7 @@ module.exports = new GearTemplate("Toxic Firecracker",
 		return `${dealDamage(targets, user, pendingDamage, false, element, adventure)} ${joinAsStatement(false, poisonedTargetNames, "is", "are", "Poisoned.")}`;
 	}
 ).setTargetingTags({ type: `random${SAFE_DELIMITER}3`, team: "foe", needsLivingTargets: true })
-	.setSidegrades("Double Firecracker", "Mercurial Firecracker")
+	.setSidegrades("Double Firecracker", "Midas's Firecracker")
 	.setModifiers({ name: "Poison", stacks: 3 })
 	.setDurability(15)
 	.setDamage(15);